remove doPostAction from post_actions and rename MessageAttachment and MessageAttachmentList components (#1956)
